feat: do not fail upgrading components

+++ lifecycle/backup_restore.sh
@@ -15,17 +15,17 @@ function check() {
   DATA_EXISTS="true"
 
   test_file='secrets.yaml'
-  if ! test -f ${DATA_PATH}/${test_file}; then
+  if ! test -f "${DATA_PATH}/${test_file}"; then
     DATA_EXISTS="false"
   fi
 }
 
 function backup() {
   echo "backing up..."
-  
+
   # uploading backup from S3
   echo "uploading storage..."
-  s3cmd --access_key=${SCW_ACCESS_KEY} --secret_key="${SCW_SECRET_KEY}" --host="https://s3.${SCW_DEFAULT_REGION}.scw.cloud" --host-bucket="https://%(bucket)s.s3.${SCW_DEFAULT_REGION}.scw.cloud" --recursive --delete-removed --force --exclude-from .s3ignore sync ${DATA_PATH}/.storage/ s3://${S3_ASSETS_BUCKET_BACKUP_PATH}/
+  s3cmd --access_key="${SCW_ACCESS_KEY}" --secret_key="${SCW_SECRET_KEY}" --host="https://s3.${SCW_DEFAULT_REGION}.scw.cloud" --host-bucket="https://%(bucket)s.s3.${SCW_DEFAULT_REGION}.scw.cloud" --recursive --delete-removed --force --exclude-from .s3ignore sync "${DATA_PATH}/.storage/" "s3://${S3_ASSETS_BUCKET_BACKUP_PATH}/"
 
   if [[ "${COPY_CONFIG}" == "true" ]]; then
     copy_configuration
@@ -36,30 +36,30 @@ function restore() {
   echo "restoring..."
 
   echo "wiping data..."
-  rm -rf ${DATA_PATH}/*
-  mkdir -p ${DATA_PATH}/.storage
+  rm -rf "${DATA_PATH:?}/"*
+  mkdir -p "${DATA_PATH}/.storage"
 
   # download backup from S3
   echo "downloading and restoring storage..."
-  s3cmd --access_key=${SCW_ACCESS_KEY} --secret_key="${SCW_SECRET_KEY}" --host="https://s3.${SCW_DEFAULT_REGION}.scw.cloud" --host-bucket="https://%(bucket)s.s3.${SCW_DEFAULT_REGION}.scw.cloud" --recursive --force sync s3://${S3_ASSETS_BUCKET_BACKUP_PATH}/ ${DATA_PATH}/.storage/
+  s3cmd --access_key="${SCW_ACCESS_KEY}" --secret_key="${SCW_SECRET_KEY}" --host="https://s3.${SCW_DEFAULT_REGION}.scw.cloud" --host-bucket="https://%(bucket)s.s3.${SCW_DEFAULT_REGION}.scw.cloud" --recursive --force sync "s3://${S3_ASSETS_BUCKET_BACKUP_PATH}/" "${DATA_PATH}/.storage/"
 
   copy_configuration
 }
 
 function copy_configuration() {
   echo "wiping current configuration data..."
-  files=$(find ${SOURCE_PATH}/common/configuration -maxdepth 1 -exec basename -a {} +)
-  for file in ${files[@]}; do
-    rm -rf ${DATA_PATH}/${file}
+  files=$(find "${SOURCE_PATH}/common/configuration" -maxdepth 1 -exec basename -a {} +)
+  for file in "${files[@]}"; do
+    rm -rf "${DATA_PATH:?}/${file}"
   done
-  files=$(find ${SOURCE_PATH}/sites/${SITE}/configuration -maxdepth 1 -exec basename -a {} +)
-  for file in ${files[@]}; do
-    rm -rf ${DATA_PATH}/${file}
+  files=$(find "${SOURCE_PATH}/sites/${SITE}/configuration" -maxdepth 1 -exec basename -a {} +)
+  for file in "${files[@]}"; do
+    rm -rf "${DATA_PATH:?}/${file}"
   done
 
   echo "copying configuration..."
-  cp -rf ${SOURCE_PATH}/common/configuration/* ${DATA_PATH}/
-  cp -rf ${SOURCE_PATH}/sites/${SITE}/configuration/* ${DATA_PATH}/
+  cp -rf "${SOURCE_PATH}/common/configuration/"* "${DATA_PATH}/"
+  cp -rf "${SOURCE_PATH}/sites/${SITE}/configuration/"* "${DATA_PATH}/"
 }
 #endregion
